refactor: remove usage of global basedenom  (#18268)
