+++ types/coin.go
@@ -881,7 +881,8 @@ func ParseCoinNormalized(coinStr string) (coin Coin, err error) {
 		return Coin{}, err
 	}
 
-	coin, _ = NormalizeDecCoin(decCoin).TruncateDecimal()
+	coin, _ = NewDecCoinFromDec(decCoin.Denom, decCoin.Amount).TruncateDecimal()
+
 	return coin, nil
 }
 
@@ -899,4 +900,21 @@ func ParseCoinsNormalized(coinStr string) (Coins, error) {
 		return Coins{}, err
 	}
 	return NormalizeCoins(coins), nil
+}
+
+// ----------------------------------------------------------------------------
+
+// NormalizeCoins normalize and truncate a list of decimal coins
+func NormalizeCoins(coins []DecCoin) Coins {
+	if coins == nil {
+		return nil
+	}
+	result := make([]Coin, 0, len(coins))
+
+	for _, coin := range coins {
+		newCoin, _ := NewDecCoinFromDec(coin.Denom, coin.Amount).TruncateDecimal()
+		result = append(result, newCoin)
+	}
+
+	return result
 }
