+++ x/distribution/keeper/delegation.go
@@ -301,9 +301,9 @@ func (k Keeper) withdrawDelegationRewards(ctx context.Context, val stakingtypes.
 	}
 
 	if finalRewards.IsZero() {
-		baseDenom, _ := sdk.GetBaseDenom()
-		if baseDenom == "" {
-			baseDenom = sdk.DefaultBondDenom
+		baseDenom, err := k.stakingKeeper.BondDenom(ctx)
+		if err != nil {
+			return nil, err
 		}
 
 		// Note, we do not call the NewCoins constructor as we do not want the zero
